requirements.txt
########################################################################################
# This section summarizes minimum dependencies and hard constraints for iTRADE.        #
#   pip install git+https://github.com/yannickberker/iTRADE                            #
# will ensure only that at least these compatible versions (or newer) are installed.   #
########################################################################################

colour-science
humanize
matplotlib
numpy
opencv-python
openpyxl
ordered-set
packaging
pandas
pyarrow
python-pptx
scikit-learn
# https://github.com/SeleniumHQ/selenium/pull/10141
selenium>=4.1.1
# >=2.8 for tf.config.experimental.enable_op_determinism
# https://github.com/tensorflow/tensorflow/issues/53771
# <2.9 for tf-keras-vis compatibility
tensorflow>=2.8,<2.9
tf-keras-vis>=0.8

########################################################################################
# The remainder of this file summarizes version numbers used for paper results. Use    #
#   pip install -r requirements.txt                                                    #
# after installing iTRADE to install these exact versions.                             #
# (Generated using pip freeze --exclude-editable >> requirements.txt)                  #
########################################################################################

absl-py==1.0.0
astroid==2.11.2
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
bandit==1.7.4
black==22.3.0
cachetools==5.0.0
certifi==2021.10.8
cffconvert==2.0.0
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colour-science==0.4.1
cryptography==36.0.2
cycler==0.11.0
Deprecated==1.2.13
dill==0.3.4
docopt==0.6.2
eradicate==2.0.0
et-xmlfile==1.1.0
flake8==4.0.1
flake8-black==0.3.2
flake8-eradicate==1.2.0
flake8-plugin-utils==1.3.2
flake8-pytest-style==1.6.0
flatbuffers==2.0
fonttools==4.31.2
gast==0.5.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h11==0.13.0
h5py==3.6.0
humanize==4.0.0
idna==3.3
imageio==2.16.1
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.1
joblib==1.1.0
jsonschema==3.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
lazy-object-proxy==1.7.1
libclang==13.0.0
lxml==4.8.0
Markdown==3.3.6
markdown2==2.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
mccabe==0.6.1
mypy==0.942
mypy-extensions==0.4.3
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
openpyxl==3.0.9
opt-einsum==3.3.0
ordered-set==4.1.0
outcome==1.1.0
packaging==21.3
pandas==1.4.2
pandas-stubs==1.2.0.56
pathspec==0.9.0
pbr==5.8.1
Pillow==9.1.0
platformdirs==2.5.1
pluggy==1.0.0
protobuf==3.20.0
py==1.11.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==2.4.0
pykwalify==1.8.0
pylama==8.3.8
pylint==2.13.4
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.1
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rpy2==3.5.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.8.0
selenium==4.1.3
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
stevedore==3.5.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tf-keras-vis==0.8.1
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
types-python-dateutil==2.8.10
types-requests==2.27.16
types-setuptools==57.4.12
types-urllib3==1.26.11
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.0
wsproto==1.1.0
XlsxWriter==3.0.3
